requirements.txt
blinker==1.7.0
broadlink==0.18.3
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
Flask==3.0.2
Flask-MQTT==1.2.1
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
paho-mqtt==1.6.1
pycparser==2.21
python-dotenv==1.0.1
Werkzeug==3.0.1
zipp==3.18.1
